dynamic environment

+++ setup.sh
@@ -1,3 +1,4 @@
-echo "testing"
-
-rm -rf ~/temp
+if [ "$ENV" = "TERMUX" ]; then
+    echo "testing"
+    rm -rf ~/temp
+fi
